refactor basic requests to utils and add a basic file request
