+++ DragonFruit.Common.Data.Tests/Requests/RequestTests.cs
@@ -3,6 +3,7 @@
 
 using DragonFruit.Common.Data.Extensions;
 using DragonFruit.Common.Data.Tests.Requests.Objects;
+using DragonFruit.Common.Data.Utils;
 using Newtonsoft.Json.Linq;
 using NUnit.Framework;
