+++ DragonFruit.Common.Data/Extensions/BasicRequestExtensions.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License. Please refer to the LICENSE file at the root of this project for details
 
 using System.Collections.Generic;
+using DragonFruit.Common.Data.Utils;
 
 namespace DragonFruit.Common.Data.Extensions
 {
